requirements.txt
asgiref==3.8.1
attrs==25.3.0
cattrs==24.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-schema-graph==3.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
pydot==3.0.4
pygraphviz==1.14
PyJWT==2.9.0
pyparsing==3.2.1
sqlparse==0.5.3
